forms-extraction-approval/container/requirements.txt
flask
azure-storage-blob
pymupdf
requests